finished connecting front-end components to one another
will fix linter errors then submit pull request
